+++ internal/cmd/cmd_test.go
@@ -35,7 +35,7 @@ import (
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-	"github.com/uber/prototool/internal/x/cmd/testdata/grpc/gen/grpcpb"
+	"github.com/uber/prototool/internal/cmd/testdata/grpc/gen/grpcpb"
 	"github.com/uber/prototool/internal/x/settings"
 	"google.golang.org/grpc"
 )
