+++ internal/cmd/testdata/bar/dep.proto
