+++ internal/cmd/testdata/compile/dep.proto
