+++ internal/cmd/testdata/compile/dep_errors.proto
