+++ internal/cmd/testdata/compile/extra_import.proto
