+++ internal/cmd/testdata/compile/not_imported.proto
