+++ internal/cmd/testdata/foo/success.proto
