+++ internal/cmd/testdata/format/bar/bar_proto2.proto.golden
