+++ internal/cmd/testdata/format/foo/foo.proto
