+++ internal/cmd/testdata/lint/file_options_required.proto
