+++ internal/cmd/testdata/lint/message_name_not_capitalized.proto
