+++ internal/cmd/testdata/lint/samedir/bar1.proto
