+++ internal/cmd/testdata/lint/samedir/foo1.proto
