+++ internal/cmd/testdata/lint/samedir/prototool.yaml
