+++ internal/cmd/testdata/lint/samedirgopkg/bar1.proto
