+++ internal/cmd/testdata/lint/samedirgopkg/foo1.proto
