+++ internal/cmd/testdata/lint/samedirgopkg/foo2.proto
