+++ internal/cmd/testdata/lint/samedirjavapkg/bar1.proto
