+++ internal/cmd/testdata/lint/samedirjavapkg/foo1.proto
