+++ internal/cmd/testdata/lint/samedirjavapkg/prototool.yaml
