+++ internal/cmd/testdata/lint/syntax_proto2.proto
